composer update

+++ database/migrations/2025_11_09_140202_add_kapasitas_to_ruang_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('ruang', function (Blueprint $table) {
-            $table->integer('kapasitas')->after('nama_ruang')->nullable();
-        });
+        if (!Schema::hasColumn('ruang', 'kapasitas')) {
+            Schema::table('ruang', function (Blueprint $table) {
+                $table->integer('kapasitas')->after('nama_ruang')->nullable();
+            });
+        }
     }
 
     /**
@@ -21,8 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('ruang', function (Blueprint $table) {
-            $table->dropColumn('kapasitas');
-        });
+        if (Schema::hasColumn('ruang', 'kapasitas')) {
+            Schema::table('ruang', function (Blueprint $table) {
+                $table->dropColumn('kapasitas');
+            });
+        }
     }
 };
